+++ packages/waggle/src/common/types.ts
@@ -52,15 +52,19 @@ export class ConnectionsManagerOptions {
   createPaths?: boolean = true
   isWaggleMobileMode?: boolean = true
   isEntryNode?: boolean = false
+  createSnapshot?: boolean = false
+  useSnapshot?: boolean = false
+  libp2pTransportClass?: any = null
 }
 
 export interface IConstructor {
   host: string
   port: number
-  agentPort: number
-  agentHost: string
+  agentPort?: number
+  agentHost?: string
   options?: Partial<ConnectionsManagerOptions>
   io: any
+  storageClass?: any // TODO: what type?
 }
 
 export interface ILibp2pStatus {