+++ packages/waggle/src/socket/events/certificates.ts
@@ -1,7 +1,7 @@
 import { EventTypesResponse } from '../constantsReponse'
 
 export const loadCertificates = (socket: SocketIO.Server, certificates: string[]) => {
-  console.log(`Sending back ${certificates.length} certificates`)
+  // console.log(`Sending back ${certificates.length} certificates`)
   socket.emit(EventTypesResponse.RESPONSE_GET_CERTIFICATES, {
     certificates
   })